Added rubles conversion and test routine
The rcheck program tests for '1R' form of price indicating
1 tsarist, Soviet, or Russian ruble.

The fixrubles.py program replace all occurrences of nR with n Rbl.

Updated price file.

The reason for this change is to prevent the currency to be dependent
on the position of the currency signature.  I was begining to get 1R
for Rubles and R 1 for South Afican Rand.  Want to avoid this
situation.

 On branch master

 Changes to be committed:
	new file:   Books20/Tools/xml/fixrubles.py
	modified:   Books20/Tools/xml/price
	new file:   Books20/Tools/xml/rcheck.sh