Added Docs/Catalogues
This software is an attempt to create a catalogue out of
of BookFile.  Initially for the Harlan J. Smith collection
but may later be used for the Books20 project.

catalogue.py is the main code location
hjs_catalogue.tex is the main LaTeX file
hjs_books.tex is the catalogue in LaTeX format. It is automatially
   built and will be removed from the repository.
bib_style contains some notes about bibliography formats

configure.ac added Docs/Catalogues/Makefile as a build target
TODO adds some notes about extending the entrylist class

 On branch master

 Changes to be committed:
	new file:   Books20/Docs/Catalogues/Makefile.am
	new file:   Books20/Docs/Catalogues/bib_style
	new file:   Books20/Docs/Catalogues/catalogue.py
	new file:   Books20/Docs/Catalogues/hjs_books.tex
	new file:   Books20/Docs/Catalogues/hjs_catalogue.tex
	modified:   Books20/TODO
	modified:   Books20/configure.ac