Add short abbreviations to search list
If the primary string passed to __get_sub_strings__ was less than the
min_length, then the primary string would not be added to the
substring list.  This meant that valid abbreviations like BA, EM, or
AJ would not be found in a search.  The string itself is now added to
the strset if the length is less than the min_length.

This clears up the problem in findjournals where journal abbreviations
like the above would not be found.

 On branch entrylist
 Changes to be committed:
	modified:   Books20/Tools/python/aabooks/lib/search.py